+++ src/app/app.ts
@@ -2,6 +2,7 @@ import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { SiteFooterComponent } from './layout/site-footer/site-footer.component';
 import { SiteHeaderComponent } from './layout/site-header/site-header.component';
+import { SeoService } from './seo/seo.service';
 import { ToastContainerComponent } from './shared/toast/toast-container.component';
 
 @Component({
@@ -10,4 +11,6 @@ import { ToastContainerComponent } from './shared/toast/toast-container.componen
   templateUrl: './app.html',
   styleUrl: './app.scss',
 })
-export class App {}
+export class App {
+  constructor(_seo: SeoService) {}
+}